+++ packages/react-native-messaging-core/ios/Sources/Models/ConversationItems.swift
@@ -2,7 +2,7 @@ import Foundation
 import NablaCore
 import NablaMessagingCore
 
-extension PaginatedList where T == ConversationItem {
+extension PaginatedList where Element == ConversationItem {
     var dictionaryRepresentation: [String: Any] {
         var result = [String: Any]()
         result["hasMore"] = hasMore