+++ packages/react-native-messaging-core/ios/Sources/Models/ConversationList.swift
@@ -2,7 +2,7 @@ import Foundation
 import NablaCore
 import NablaMessagingCore
 
-extension PaginatedList where T == Conversation {
+extension PaginatedList where Element == Conversation {
     var dictionaryRepresentation: [String: Any] {
         [
             "conversations": elements.map(\.dictionaryRepresentation),