+++ packages/react-native-scheduling/ios/Sources/NablaAppointmentListViewManager/NablaAppointmentListViewManager.m
@@ -2,4 +2,6 @@
 
 @interface RCT_EXTERN_MODULE(NablaAppointmentListViewManager, RCTViewManager)
 
+RCT_EXPORT_VIEW_PROPERTY(onAppointmentSelected, RCTDirectEventBlock)
+
 @end