+++ packages/react-native-scheduling/ios/Sources/NablaSchedulingClient/NablaSchedulingClientModule.m
@@ -5,6 +5,4 @@
 RCT_EXTERN_METHOD(initializeSchedulingModule: (RCTPromiseResolveBlock)resolve
                                     rejecter: (RCTPromiseRejectBlock)reject)
 
-RCT_EXTERN_METHOD(openScheduleAppointmentScreen)
-
 @end