+++ packages/react-native-scheduling/ios/Sources/NablaSchedulingClient/NablaSchedulingClientModule.swift
@@ -12,17 +12,6 @@ final class NablaSchedulingClientModule: NSObject {
         resolver(NSNull())
     }
 
-    @objc(openScheduleAppointmentScreen)
-    func openScheduleAppointmentScreen() {
-        DispatchQueue.main.async {
-            guard let rootViewController = UIApplication.shared.delegate?.window??.rootViewController else {
-                CoreLogger.sharedInstance.error(message: "Missing Application window rootViewController")
-                return
-            }
-            NablaClient.shared.scheduling.views.presentScheduleAppointmentViewController(from: rootViewController)
-        }
-    }
-
     // MARK: - Overridden
 
     @objc(requiresMainQueueSetup)