+++ packages/react-native-scheduling/ios/Sources/NablaSchedulingUI/NablaSchedulingUIModule.m
@@ -1,0 +1,9 @@
+#import <React/RCTBridgeModule.h>
+
+@interface RCT_EXTERN_MODULE(NablaSchedulingUIModule, NSObject)
+
+RCT_EXTERN_METHOD(navigateToScheduleAppointmentScreen)
+RCT_EXTERN_METHOD(navigateToAppointmentDetailScreen: (NSString *)appointmentId
+                                           callback: (RCTResponseSenderBlock)callback)
+
+@end