+++ packages/react-native-scheduling/src/NablaSchedulingClient/NablaSchedulingClient.ts
@@ -12,8 +12,4 @@ export class NablaSchedulingClient {
   public static async initializeSchedulingModule() {
     await nablaSchedulingClientModule.initializeSchedulingModule();
   }
-
-  public static openScheduleAppointmentScreen() {
-    nablaSchedulingClientModule.openScheduleAppointmentScreen()
-  }
 }