+++ packages/react-native-scheduling/src/NablaSchedulingUI/NablaSchedulingUIModule.ts
@@ -1,4 +1,5 @@
 import { NativeModule, NativeModules, Platform } from 'react-native';
+import { Callback } from '@nabla/react-native-core/lib/internal';
 
 const LINKING_ERROR =
   `The package '@nabla/react-native-scheduling' doesn't seem to be linked. Make sure: \n\n` +
@@ -6,13 +7,16 @@ const LINKING_ERROR =
   '- You rebuilt the app after installing the package\n' +
   '- You are not using Expo managed workflow\n';
 
-interface NablaSchedulingClientModule extends NativeModule {
-  initializeSchedulingModule(): Promise<void>;
+interface NablaSchedulingUIModule extends NativeModule {
+  navigateToScheduleAppointmentScreen(): void;
+  navigateToAppointmentDetailScreen(
+    appoinmentId: string,
+    callbacl: Callback<void>): void;
 }
 
-export const nablaSchedulingClientModule: NablaSchedulingClientModule =
-  NativeModules.NablaSchedulingClientModule
-    ? NativeModules.NablaSchedulingClientModule
+export const nablaSchedulingUIModule: NablaSchedulingUIModule =
+  NativeModules.NablaSchedulingUIModule
+    ? NativeModules.NablaSchedulingUIModule
     : new Proxy(
         {},
         {