+++ packages/react-native-scheduling/src/index.ts
@@ -1,2 +1,4 @@
 export { NablaSchedulingClient } from './NablaSchedulingClient/NablaSchedulingClient';
+export { NablaSchedulingUI } from './NablaSchedulingUI/NablaSchedulingUI';
 export { NablaAppointmentListView } from './NablaAppointmentListView/NablaAppointmentListView'
+export * from './types';